pyproject.toml
[project]
name = "conversational-ai-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gtts>=2.5.4",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "openai>=1.107.1",
    "python-dotenv>=1.1.1",
    "speechrecognition>=3.14.3",
    "streamlit>=1.49.1",
]
